Execute all the logic inside the after-build event
This commit removes the need to use the before-build-all event for
initializing the state of the plugin. Instead, everything is initialized
in the after-build event, along with the minifying code.

This removes the chance of inconsistencies due to the global state used,
and possible errors if events are fired in the wrong order.